// src/Framework/SPPC.Tadbir.Web.New/ClientApp/src/app/finance/models/fullaccount.ts
import { IEntity } from "./IEntity";
import { AccountItemBrief } from "./accountItemBrief";

// ------------------------------------------------------------------------------
// <auto-generated>
//     This code was generated by a tool.
//     Runtime Version: 1.1.282
//     Template Version: 1.0
//     Generation Date: 05/09/2018 09:46:19 ق.ظ
//  
//     Changes to this file may cause incorrect behavior and will be lost if
//     the code is regenerated.
// </auto-generated>
// ------------------------------------------------------------------------------


export interface FullAccount extends IEntity{
    account: AccountItemBrief;
    detailAccount: AccountItemBrief;
    costCenter: AccountItemBrief;
    project: AccountItemBrief;
}